Share todo summary item projection helper
