+++ loopx/control_plane/todos/claim_visibility.py
@@ -5,106 +5,18 @@ from typing import Any
 from .contract import (
     TODO_TASK_CLASS_ADVANCEMENT,
     TODO_TASK_CLASS_MONITOR,
-    normalize_required_write_scopes,
     normalize_todo_claimed_by,
-    normalize_todo_decision_scope,
-    normalize_todo_required_decision_scopes,
-    normalize_todo_task_class,
 )
 from .projection import (
     todo_claimed_visibility_items,
     todo_item_is_actionable_open,
+    todo_item_task_class,
     todo_projection_sort_key,
 )
+from .summary_item import compact_todo_summary_item
 
 
 TODO_AGENT_CLAIM_SCOPE_SCHEMA_VERSION = "agent_claim_scope_v0"
-
-
-def _todo_task_class(item: dict[str, Any]) -> str:
-    text = " ".join(
-        str(value or "")
-        for value in (item.get("title"), item.get("text"))
-        if str(value or "").strip()
-    )
-    return normalize_todo_task_class(
-        item.get("task_class"),
-        text=text,
-        action_kind=item.get("action_kind"),
-    )
-
-
-def _compact_claim_visibility_item(
-    item: dict[str, Any],
-    *,
-    text: str,
-) -> dict[str, Any]:
-    compact: dict[str, Any] = {
-        "index": item.get("index"),
-        "text": text,
-    }
-    for key in (
-        "schema_version",
-        "todo_id",
-        "role",
-        "status",
-        "priority",
-        "title",
-        "archive_state",
-        "source_section",
-        "task_class",
-        "action_kind",
-        "required_write_scopes",
-        "required_capabilities",
-        "target_capabilities",
-        "decision_scope",
-        "required_decision_scopes",
-        "claimed_by",
-        "blocks_agent",
-        "unblocks_todo_id",
-        "resume_when",
-        "resume_condition",
-        "resume_ready",
-        "no_followup",
-        "successor_todo_ids",
-        "target_key",
-        "cadence",
-        "next_due_at",
-        "expires_at",
-        "last_checked_at",
-        "result_hash",
-        "consecutive_no_change",
-        "material_change",
-        "max_no_change_before_replan",
-        "route_continuation_replan_required",
-        "route_continuation_reason",
-        "route_id",
-        "route_key",
-        "completed_at",
-        "updated_at",
-        "superseded_by",
-    ):
-        if item.get(key) is not None:
-            compact[key] = item.get(key)
-    required_write_scopes = normalize_required_write_scopes(compact.get("required_write_scopes"))
-    if required_write_scopes:
-        compact["required_write_scopes"] = required_write_scopes
-    else:
-        compact.pop("required_write_scopes", None)
-    decision_scope = normalize_todo_decision_scope(compact.get("decision_scope"))
-    if decision_scope:
-        compact["decision_scope"] = decision_scope
-    else:
-        compact.pop("decision_scope", None)
-    required_decision_scopes = normalize_todo_required_decision_scopes(
-        compact.get("required_decision_scopes")
-    )
-    if required_decision_scopes:
-        compact["required_decision_scopes"] = required_decision_scopes
-    else:
-        compact.pop("required_decision_scopes", None)
-    compact["task_class"] = _todo_task_class(compact)
-    return compact
 
 
 def _compact_items(
@@ -113,7 +25,7 @@ def _compact_items(
     limit: int,
 ) -> list[dict[str, Any]]:
     return [
-        _compact_claim_visibility_item(item, text=str(item.get("text") or "").strip())
+        compact_todo_summary_item(item, text=str(item.get("text") or "").strip())
         for item in items[:limit]
     ]
 
@@ -194,13 +106,13 @@ def build_todo_claim_visibility_lanes(
         item
         for item in claimed_items
         if todo_item_is_actionable_open(item)
-        if _todo_task_class(item) == TODO_TASK_CLASS_ADVANCEMENT
+        if todo_item_task_class(item) == TODO_TASK_CLASS_ADVANCEMENT
     ]
     claimed_monitor_items = [
         item
         for item in claimed_items
         if todo_item_is_actionable_open(item)
-        if _todo_task_class(item) == TODO_TASK_CLASS_MONITOR
+        if todo_item_task_class(item) == TODO_TASK_CLASS_MONITOR
     ]
 
     lanes: dict[str, Any] = {
@@ -251,13 +163,13 @@ def build_todo_claim_visibility_lanes(
             item
             for item in current_agent_items
             if todo_item_is_actionable_open(item)
-            if _todo_task_class(item) == TODO_TASK_CLASS_ADVANCEMENT
+            if todo_item_task_class(item) == TODO_TASK_CLASS_ADVANCEMENT
         ]
         current_agent_monitor_items = [
             item
             for item in current_agent_items
             if todo_item_is_actionable_open(item)
-            if _todo_task_class(item) == TODO_TASK_CLASS_MONITOR
+            if todo_item_task_class(item) == TODO_TASK_CLASS_MONITOR
         ]
         lanes.update(
             {
